Fix inconsistent deliveryId bug on server.

+++ src/server/cftp/models/Request.go
@@ -9,5 +9,5 @@ type Request struct {
 
 type MetaData struct {
 	SenderAddress string
-	RequestId     int
+	RequestId     int64
 }
